refactor (repository): replace int type with int64 for ids

+++ internal/domain/payment_repository.go
@@ -4,9 +4,9 @@ import "context"
 
 type PaymentRepository interface {
 	Create(ctx context.Context, payment *Payment) error
-	List(ctx context.Context, userId int) ([]*Payment, error)
-	Find(ctx context.Context, id int) (*Payment, error)
+	List(ctx context.Context, userId int64) ([]*Payment, error)
+	Find(ctx context.Context, id int64) (*Payment, error)
 	Update(ctx context.Context, payment *Payment) error
-	Delete(ctx context.Context, id int) error
-	Count(ctx context.Context, userId int) (int, error)
+	Delete(ctx context.Context, id int64) error
+	Count(ctx context.Context, userId int64) (int, error)
 }
